sorveteria-pos/agent/requirements.txt
fastapi>=0.110
uvicorn>=0.27
python-dotenv>=1.0
pyserial>=3.5
python-escpos>=3.1
pywin32>=306